add VoiceActivity.kt
